move DashboardsBar to /components
