+++ src/components/DashboardsBar/Chip.js
@@ -5,8 +5,8 @@ import debounce from 'lodash/debounce'
 import PropTypes from 'prop-types'
 import React from 'react'
 import { Link } from 'react-router-dom'
-import { apiPostDataStatistics } from '../../../api/dataStatistics'
-import { useCacheableSection } from '../../../modules/useCacheableSection'
+import { apiPostDataStatistics } from '../../api/dataStatistics'
+import { useCacheableSection } from '../../modules/useCacheableSection'
 import { OfflineSaved } from './assets/icons'
 import classes from './styles/Chip.module.css'
