+++ src/components/DashboardsBar/ClearButton.js
