+++ src/components/DashboardsBar/Content.js
@@ -6,9 +6,9 @@ import PropTypes from 'prop-types'
 import React, { useState } from 'react'
 import { connect } from 'react-redux'
 import { Redirect, withRouter } from 'react-router-dom'
-import { sGetAllDashboards } from '../../../reducers/dashboards'
-import { sGetDashboardsFilter } from '../../../reducers/dashboardsFilter'
-import { sGetSelectedId } from '../../../reducers/selected'
+import { sGetAllDashboards } from '../../reducers/dashboards'
+import { sGetDashboardsFilter } from '../../reducers/dashboardsFilter'
+import { sGetSelectedId } from '../../reducers/selected'
 import Chip from './Chip'
 import Filter from './Filter'
 import { getFilteredDashboards } from './getFilteredDashboards'
