+++ src/components/DashboardsBar/DashboardsBar.js
@@ -8,10 +8,10 @@ import React, {
     createRef,
 } from 'react'
 import { connect } from 'react-redux'
-import { acSetControlBarUserRows } from '../../../actions/controlBar'
-import { apiPostControlBarRows } from '../../../api/controlBar'
-import { useWindowDimensions } from '../../../components/WindowDimensionsProvider'
-import { sGetControlBarUserRows } from '../../../reducers/controlBar'
+import { acSetControlBarUserRows } from '../../actions/controlBar'
+import { apiPostControlBarRows } from '../../api/controlBar'
+import { useWindowDimensions } from '../../components/WindowDimensionsProvider'
+import { sGetControlBarUserRows } from '../../reducers/controlBar'
 import Content from './Content'
 import DragHandle from './DragHandle'
 import { getRowsFromHeight } from './getRowsFromHeight'
