+++ src/components/DashboardsBar/__tests__/Chip.spec.js
