+++ src/components/DashboardsBar/__tests__/ClearButton.spec.js
