+++ src/components/DashboardsBar/__tests__/DashboardsBar.spec.js
