+++ src/components/DashboardsBar/__tests__/getRowsFromHeight.js
