+++ src/components/DashboardsBar/assets/Clear.js
