+++ src/components/DashboardsBar/assets/icons.js
