+++ src/components/DashboardsBar/getRowsFromHeight.js
