+++ src/pages/view/ViewDashboard.js
@@ -14,6 +14,7 @@ import {
 } from '../../actions/selected'
 import { apiPostDataStatistics } from '../../api/dataStatistics'
 import DashboardContainer from '../../components/DashboardContainer'
+import DashboardsBar from '../../components/DashboardsBar/DashboardsBar'
 import LoadingMask from '../../components/LoadingMask'
 import Notice from '../../components/Notice'
 import { setHeaderbarVisible } from '../../modules/setHeaderbarVisible'
@@ -21,7 +22,6 @@ import { useCacheableSection } from '../../modules/useCacheableSection'
 import { sGetDashboardById } from '../../reducers/dashboards'
 import { sGetPassiveViewRegistered } from '../../reducers/passiveViewRegistered'
 import { sGetSelectedId } from '../../reducers/selected'
-import DashboardsBar from './DashboardsBar/DashboardsBar'
 import FilterBar from './FilterBar/FilterBar'
 import ItemGrid from './ItemGrid'
 import classes from './styles/ViewDashboard.module.css'
