+++ src/pages/view/index.js
@@ -1,4 +1,4 @@
+import DashboardsBar from '../../components/DashboardsBar/DashboardsBar'
 import ViewDashboard from './CacheableViewDashboard'
-import DashboardsBar from './DashboardsBar/DashboardsBar'
 
 export { ViewDashboard, DashboardsBar }
